Fix very low cost issue and add Share functionality
Add the ability to easily share stats about your metro usage and amount saved using the tool. Also, fixed an issue where if you had a pass and it didn't cover the cost of the trip fully, the algorithm assumed the price for the trip was the difference. Eg: $2.25 pass, trip costs $2.30, algorithm says price is $.05.